Fix and refactor vhserver_player_info.py
After noticing that running the previous version through the systemd service caused the script to continuously print the player database inside the "output log file", I decided to change the execution logic. Now, the default behaviour is to list the player database, while launching the script with the "--daemon" parameter executes the background process.

ChatGPT helped with the code refactoring and formatting, so now the code looks cleaner (and it has classes now).
I've also removed the "player_serial" and replaced it with the "player_steamID", and added a "player_game_sessions" variable to show how often the player connects to the server (I still have to understand how to consistently ascertain when a specific player disconnects from the server).